[P4ADEV-1895] refactor sidebar and debt types components to use new super admin utility
