+++ src/utils/index.ts
@@ -6,6 +6,7 @@ import { Api } from '../../generated/apiClient';
 import { FileshareApi } from '../../generated/fileshare/fileshareClient';
 import storage from './storage';
 import notify from './notify';
+import roles from './roles';
 
 export default {
   apiClient: new Api({ baseURL: config.baseURL, timeout: config.apiTimeout }),
@@ -18,5 +19,6 @@ export default {
   sidemenu,
   style,
   storage,
-  notify
+  notify,
+  roles
 };
